pyproject.toml
[tool.poetry]
name = "testtaskwb"
version = "0.9.0"
description = ""
authors = ["GutesTod <jefflol003@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.4.1"
sqlalchemy = "^2.0.28"
python-dotenv = "^1.0.1"
apscheduler = "^3.10.4"
sqlalchemy-utils = "^0.41.1"
psycopg = {extras = ["binary"], version = "^3.1.18"}
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
